Refactor network packet class organization and update imports
